requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
networkx>=3.0
numpy>=1.24.0
tiktoken>=0.5.0

# Data processing
datasets>=2.14.0
pandas>=2.0.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Graph processing and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
jsonlines>=3.1.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.24.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3